Add: server.js & app.js 파일 분리
