+++ server.js
@@ -1,7 +1,4 @@
 const http = require('http');
-const express = require('express');
-const cors = require('cors');
-const morgan = require('morgan');
 
 const dotenv = require('dotenv');
 dotenv.config();
@@ -9,22 +6,15 @@ console.log('DB_HOST:', process.env.DB_HOST);
 console.log('DB_USER:', process.env.DB_USER);
 console.log('DB_PASS:', process.env.DB_PASS);
 
-const routes = require('./routes');
+const createApp = require('./app');
 
-var corsOptions = {
-  origin: '*',
-  optionsSuccessStatus: 200,
+const startServer = async () => {
+  const app = createApp();
+  const PORT = process.env.PORT;
+
+  app.listen(PORT, () => {
+    console.log(`server is listening on PORT ${PORT}`);
+  });
 };
 
-const app = express();
-
-app.use(cors(corsOptions));
-app.use(express.json());
-app.use(routes);
-app.use(morgan('combined'));
-
-const server = http.createServer(app);
-const PORT = process.env.PORT || 10010;
-server.listen(PORT, () => {
-  console.log(`server start : http://localhost:${PORT}/`);
-});
+startServer();
